extract seeds runner file from seed class file
