+++ db/seeds.rb
@@ -1,37 +1,3 @@
-require 'csv'
-require './app/models/station'
-require './app/models/trip'
-
-class Seed
-
-  def self.start
-    options = {headers: true, header_converters: :symbol, converters: :numeric}
-
-    CSV.foreach("./db/csv/station.csv", options ) do |row|
-      next if Station.exists?(row[:id])
-      convert_date(row, :installation_date)
-      Station.create!(row.to_hash)
-    end
-
-    CSV.foreach("./db/csv/trip_fixture.csv", options ) do |row|
-      next if Trip.exists?(row[:id])
-      next unless Station.exists?(row[:start_station_id])
-      next unless Station.exists?(row[:end_station_id])
-      convert_date(row, :start_date)
-      convert_date(row, :end_date)
-      row.delete(:start_station_name)
-      row.delete(:end_station_name)
-      Trip.create(row.to_hash)
-
-    end
-
-  end
-
-  def self.convert_date(row, field)
-    m, d, y = row[field].split('/')
-    row[field] = [d, m, y].join('/')
-  end
-
-end
+require_relative 'seed'
 
 Seed.start
